# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.CheckinRequest:
    properties:
      action:
        enum:
        - board
        - exit
        - school_recv
        type: string
      location:
        $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LocationDTO'
      student_id:
        type: string
    required:
    - action
    - student_id
    type: object
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.CheckinResponse:
    properties:
      action:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LocationDTO'
      student_id:
        type: string
      timestamp:
        type: string
      trip_id:
        type: string
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.EndTripRequest:
    properties:
      end_location:
        $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LocationDTO'
    type: object
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ListTripsResponse:
    properties:
      total:
        type: integer
      trips:
        items:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.TripResponse'
        type: array
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LocationDTO:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.StartTripRequest:
    properties:
      route_id:
        type: string
    required:
    - route_id
    type: object
...
  github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.TripResponse:
    properties:
      driver_id:
        type: string
      end_location:
        $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LocationDTO'
      ended_at:
        type: string
      id:
        type: string
      route_id:
        type: string
      start_location:
        $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.LocationDTO'
      started_at:
        type: string
      status:
        type: string
      vehicle_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API documentation for the gateway service of the School Tracking System.
  title: Gateway Service API
  version: "1.0"
paths:
...
  /trips:
    get:
      description: Paginated list of trips, filterable by status or route
      parameters:
      - description: Limit results (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset results (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by Route ID
        in: query
        name: route_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ListTripsResponse'
      security:
      - BearerAuth: []
      summary: List trips
      tags:
      - trips
    post:
      consumes:
      - application/json
      description: Create a new trip instance based on a route ID
      parameters:
      - description: Start Trip Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.StartTripRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.TripResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start a new trip
      tags:
      - trips
  /trips/{id}:
    get:
      description: Retrieve full trip data by UUID
      parameters:
      - description: Trip UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.TripResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get trip details
      tags:
      - trips
  /trips/{id}/checkins:
    get:
      description: Retrieve all student check-in events for a specific trip
      parameters:
      - description: Trip UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.CheckinResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get trip check-ins
      tags:
      - trips
    post:
      consumes:
      - application/json
      description: Register a boarding/exiting event for a student on a trip
      parameters:
      - description: Trip UUID
        in: path
        name: id
        required: true
        type: string
      - description: Checkin Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.CheckinRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.CheckinResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Student check-in
      tags:
      - trips
  /trips/{id}/end:
    put:
      consumes:
      - application/json
      description: Mark an in-progress trip as completed
      parameters:
      - description: Trip UUID
        in: path
        name: id
        required: true
        type: string
      - description: End Trip Request
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.EndTripRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.TripResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: End a trip
      tags:
      - trips
  /trips/active:
    get:
      description: Retrieve all trips currently in progress
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_fercho_school-tracking_services_gateway_internal_infrastructure_api_dtos.TripResponse'
            type: array
      security:
      - BearerAuth: []
      summary: List active trips
      tags:
      - trips
securityDefinitions:
  BearerAuth:
    description: 'JWT token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"